ms4-prevision-eau/requirements.txt
fastapi
uvicorn
pandas
numpy
prophet
torch
psycopg2-binary
pydantic-settings
scikit-learn
pytest
requests
urllib3